[meta] Stop publishing development files
This package is currently being published with unnecessary files, such
as `.eslintrc` and the `test/` directory.

This removes the following files:

- .eslintrc
- .github/FUNDING.yml
- .nycrc
- test/index.js

You can see the difference by comparing the results of `npm pack
--dry-run`.

It does this by replacing `.npmignore`, a blocklist, with [the `files`
property in package.json][0], an allowlist.

[0]: https://docs.npmjs.com/cli/v8/configuring-npm/package-json/#files